requirements.txt
# Web app / DB
Flask
pymongo
python-dotenv

# LLM + memory pipeline
openai
google-genai
mem0ai
pydantic
tiktoken

# Data / analytics
numpy
pandas
scipy
scikit-learn
statsmodels
tqdm

# NLP / task generation
nltk
spacy
thefuzz
convokit

# Entropy / LM scripts
torch
transformers
datasets
